Reuse some checks
