+++ src/Common/Coverage.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#if defined(SANITIZE_COVERAGE)
+void dumpCoverage();
+#endif
