+++ src/Common/EnvironmentChecks.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#if !defined(USE_MUSL)
+void checkHarmfulEnvironmentVariables(char ** argv);
+#endif
